Draw simple rectangles as esp instead
Use simple rectangles instead of hitboxes.
Hitboxes were drawn over the playermodel making it hard to see
if the player is actually behind a wall or not. rectangles r simple
and visiually more clear while also reducing code in comparison to
the hitbox method and using just as little resources if not even less
than the hitbox method.

removed all of the code related to the hitbox method.

Client.h:
[L.407]
initiate DrawESP function.

Client.cpp:
[L.480]
upon joining a team set followedplayerid to localplayer's id. If we
dont determine followedplayerid early on it will cause a crash when
we initially join as spectator as DrawESP needs it that early on.

Client_Draw.cpp:
[L.73]
setup setting for turning esp rectangles on or off ingame.
[L.354-378]
DrawESP Function for drawing the esp rectangle for a given player.
[L.399-404]
include DrawESP in for loop from DrawPUBOVL. it will be drawn
alongside pubovl playernames when cg_spectatorESP setting is on.
(L.400-401) dont draw the rectangle from the player you spectate.
otherwise it would obstruct ur view. this is where we need
followedplayerid and why we need Client.cpp:[L.480].

Preferences.as:
[L.663]
rename dd_spectatorESP to cg_spectatorESP. the entire code now has
pretty much nothing to do with doctor dank. adding more prefixes such
as "dd_" may cause more confusion and players may not want to remember
all of them. i guess "cg_" stands for control graphics and since the
esp rectangles r a graphical feature they should be included to this
prefix. cg_ is also familiar to most players as its used for most of
the other settings ingame which r of graphical nature.